AdministratorPaymentsTemplate renamed to AdministratorViewTemplate. Now component is reusable. It take two props: configuration object and component.
