+++ src/components/templates/AdministratorNavigation/AdministratorNavigation.js
@@ -3,12 +3,13 @@ import { BrowserRouter as Router, Switch, Route } from 'react-router-dom';
 
 import { ThemeProvider } from 'styled-components';
 import { theme } from '../../../styles/styles';
-import AdministratorPaymentsTemplate from '../AdministratorPaymentsTemplate/AdministratorPaymentsTemplate';
+import AdministratorViewTemplate from '../AdministratorViewTemplate/AdministratorViewTemplate';
 import GridContainerNavigationWrapper from '../../../styles/stylesContainer/GridContainerNavigationWrapper';
 import Navigation from '../../Organisms/Navigation/Navigation';
 import { administratorNavigation } from '../../../helpers/navigationNamesAndRoutes/administratorNavigation';
 import { AnimalsForAdoption as Title } from '../../Atoms/HeaderAnimalsForAdoption/HeaderAnimalsForAdoption';
-import { administratorPaymentConfig} from '../../../Config/administratorViewConfigFile';
+import { administratorPaymentConfig } from '../../../Config/administratorViewConfigFile';
+import AdministratorPayment from '../../Organisms/AdministratorPayment/AdministratorPayment';
 
 const NavigationTemplate = ({ children }) => {
   return (
@@ -26,9 +27,7 @@ function AdministratorNavigation() {
     <ThemeProvider theme={theme}>
       <Router>
         <NavigationTemplate>
-          <Title
-            style={{ textAlign: 'center', marginTop: '50px' }}
-          >
+          <Title style={{ textAlign: 'center', marginTop: '50px' }}>
             Panel Administratora
           </Title>
           <Switch>
@@ -39,7 +38,10 @@ function AdministratorNavigation() {
 
             <Route path="/payments">
               {/* <Wrapper> */}
-              <AdministratorPaymentsTemplate administratorConfig={administratorPaymentConfig} />
+              <AdministratorViewTemplate
+                administratorConfig={administratorPaymentConfig}
+                componentName={<AdministratorPayment />}
+              ></AdministratorViewTemplate>
               {/* </Wrapper> */}
             </Route>
           </Switch>
